# objectified-rest/openapi/openapi.yaml
openapi: 3.2.0
info:
  title: Objectified REST API
  description: REST API for the Objectified platform. CRUD scaffolding from database
    design. Services are used by the platform and can be exposed externally; internal
    use is via a private API key.
  version: 1.0.17
paths:
...
  /v1/users:
    get:
      tags:
      - Users
      summary: List users (admin)
      description: List all accounts. **Admin only** — requires a valid JWT with ``is_admin=true``,
        an account that is an administrator in at least one tenant, or a valid internal
        API key. Soft-deleted accounts are excluded by default; pass ``include_deleted=true``
        to include them.
      operationId: list_users_v1_users_get
      parameters:
      - name: include_deleted
        in: query
        required: false
        schema:
          type: boolean
          description: Include soft-deleted accounts
          default: false
          title: Include Deleted
        description: Include soft-deleted accounts
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountSchema'
                title: Response List Users V1 Users Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Users
      summary: Create user (Sign Up)
      description: Register a new user account. Email must be unique.
      operationId: create_user_v1_users_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/tenants:
    get:
      tags:
      - Tenants
      summary: List tenants
      description: List all tenants. Soft-deleted tenants are excluded by default.
      operationId: list_tenants_v1_tenants_get
      parameters:
      - name: include_deleted
        in: query
        required: false
        schema:
          type: boolean
          description: Include soft-deleted tenants
          default: false
          title: Include Deleted
        description: Include soft-deleted tenants
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TenantSchema'
                title: Response List Tenants V1 Tenants Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Tenants
      summary: Create tenant
      description: Create a new tenant. Slug must be unique and URL-safe (lowercase
        alphanumeric with hyphens). Requires authentication. The authenticated user
        is assigned as an administrator of the new tenant.
      operationId: create_tenant_v1_tenants_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/tenants/{tenant_id}/members:
    get:
      tags:
      - Tenants
      summary: List tenant members
      description: List all active members (tenant_account rows) for a tenant.
      operationId: list_tenant_members_v1_tenants__tenant_id__members_get
      parameters:
      - name: tenant_id
        in: path
        required: true
        schema:
          type: string
          title: Tenant Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TenantAccountSchema'
                title: Response List Tenant Members V1 Tenants  Tenant Id  Members
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Tenants
      summary: Add tenant member
      description: Add an account to a tenant with a given access level. The account
        can be identified by ``account_id`` (UUID) or ``email``. If both are provided,
        ``account_id`` takes precedence.
      operationId: add_tenant_member_v1_tenants__tenant_id__members_post
      parameters:
      - name: tenant_id
        in: path
        required: true
        schema:
          type: string
          title: Tenant Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantAccountCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantAccountSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/tenants/{tenant_id}/administrators:
    get:
      tags:
      - Tenants
      summary: List tenant administrators
      description: List all active members with access_level=administrator for a tenant.
      operationId: list_tenant_administrators_v1_tenants__tenant_id__administrators_get
      parameters:
      - name: tenant_id
        in: path
        required: true
        schema:
          type: string
          title: Tenant Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TenantAccountSchema'
                title: Response List Tenant Administrators V1 Tenants  Tenant Id  Administrators
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Tenants
      summary: Add tenant administrator
      description: Add an account to a tenant with the ``administrator`` access level,
        or promote an existing member to administrator. The account can be identified
        by ``account_id`` (UUID) or ``email``. If both are provided, ``account_id``
        takes precedence. **Admin only.**
      operationId: add_tenant_administrator_v1_tenants__tenant_id__administrators_post
      parameters:
      - name: tenant_id
        in: path
        required: true
        schema:
          type: string
          title: Tenant Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantAdministratorCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantAccountSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AccountCreate:
      properties:
        name:
          type: string
          title: Name
        email:
          type: string
          title: Email
        password:
          type: string
          title: Password
        metadata:
          additionalProperties: true
          type: object
          title: Metadata
      type: object
      required:
      - name
      - email
      - password
      title: AccountCreate
      description: 'Create payload for objectified.account (password required).


        ``verified`` and ``enabled`` are intentionally omitted: new sign-ups are

        always created with ``verified=False`` and ``enabled=True`` on the server

        side to prevent callers from self-verifying or disabling their account.'
...
    AccountSchema:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        email:
          type: string
          title: Email
        verified:
          type: boolean
          title: Verified
          default: false
        enabled:
          type: boolean
          title: Enabled
          default: true
        metadata:
          additionalProperties: true
          type: object
          title: Metadata
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
        deleted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deleted At
      type: object
      required:
      - id
      - name
      - email
      - created_at
      title: AccountSchema
      description: Response schema for objectified.account.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    TenantAccessLevel:
      type: string
      enum:
      - member
      - administrator
      title: TenantAccessLevel
      description: Enum for objectified.tenant_access_level.
    TenantAccountCreate:
      properties:
        tenant_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Tenant Id
        account_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Account Id
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        access_level:
          $ref: '#/components/schemas/TenantAccessLevel'
          default: member
        enabled:
          type: boolean
          title: Enabled
          default: true
      type: object
      title: TenantAccountCreate
      description: 'Create payload for objectified.tenant_account.


        Either ``account_id`` (UUID) or ``email`` must be provided to identify the

        account to add.  If both are supplied, ``account_id`` takes precedence.

        ``tenant_id`` is optional in the body — it is validated against the path

        parameter in the route handler.'
    TenantAccountSchema:
      properties:
        id:
          type: string
          title: Id
        tenant_id:
          type: string
          title: Tenant Id
        account_id:
          type: string
          title: Account Id
        access_level:
          $ref: '#/components/schemas/TenantAccessLevel'
          default: member
        enabled:
          type: boolean
          title: Enabled
          default: true
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
        deleted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deleted At
      type: object
      required:
      - id
      - tenant_id
      - account_id
      - created_at
      title: TenantAccountSchema
      description: Response schema for objectified.tenant_account.
...
    TenantAdministratorCreate:
      properties:
        tenant_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Tenant Id
        account_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Account Id
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        enabled:
          type: boolean
          title: Enabled
          default: true
      additionalProperties: false
      type: object
      title: TenantAdministratorCreate
      description: 'Request body for POST /v1/tenants/{id}/administrators.


        Dedicated schema that intentionally omits ``access_level`` — the endpoint

        always assigns ``administrator`` and the field is not meaningful here.


        Either ``account_id`` (UUID) or ``email`` must be provided to identify the

        account.  If both are supplied, ``account_id`` takes precedence.

        ``tenant_id`` is optional in the body — it is validated against the path

        parameter in the route handler.'
    TenantCreate:
      properties:
        name:
          type: string
          title: Name
        description:
          type: string
          title: Description
          default: ''
        slug:
          type: string
          pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
          title: Slug
          description: 'URL-safe tenant identifier: lowercase alphanumeric segments
            separated by single hyphens (e.g. ''my-tenant'', ''acme'', ''acme-corp-2'').
            Must not start or end with a hyphen.'
        enabled:
          type: boolean
          title: Enabled
          default: true
        metadata:
          additionalProperties: true
          type: object
          title: Metadata
      type: object
      required:
      - name
      - slug
      title: TenantCreate
      description: Create payload for objectified.tenant.
    TenantSchema:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        description:
          type: string
          title: Description
        slug:
          type: string
          title: Slug
        enabled:
          type: boolean
          title: Enabled
          default: true
        metadata:
          additionalProperties: true
          type: object
          title: Metadata
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
        deleted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deleted At
      type: object
      required:
      - id
      - name
      - description
      - slug
      - created_at
      title: TenantSchema
      description: Response schema for objectified.tenant.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT token from NextAuth (Authorization: Bearer <token>)'
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for tenant-scoped access (alternative to JWT)